Update branches, add fahrenheit_to_celcius_converter
finished up the chapter on control flow. the fahrenheit to celcius converter is part of the summary; there are a couple other projects that will come from there. I'm still trying to wrap my head around some of the stuff Rust is a stickler on (like unused variables and data types)